config: settle pre-1.0 naming + type consistency (BREAKING)
Small, targeted pre-1.0 Config cleanup (a survey found the naming mostly
already consistent — quic-zig's allow_/reveal_/insecure_ prefixes are
semantically justified and left as-is):

- Rename the two http3-zig on/off toggles that lacked the enable_ prefix:
  qpack_huffman -> enable_qpack_huffman, open_qpack_streams ->
  enable_qpack_streams, matching enable_connect_protocol / enable_datagram /
  enable_webtransport.
- Unify max_field_section_size to ?u64 across session.Config, ProductionOptions,
  the curated Client.Config / Server.Config, and the message codec — it was
  ?usize while the wire type in Settings is ?u64. Removes the @intCast dance;
  source-compatible for integer-literal callers on 64-bit targets.

Also records this sprint's DoS-cap hardening (the four capped session maps)
in the CHANGELOG and documents the settled Config convention in
docs/API_STABILITY.md.

+++ src/message.zig
@@ -38,12 +38,12 @@ pub const Kind = enum {
 };
 
 pub const EncodeOptions = struct {
-    max_field_section_size: ?usize = null,
+    max_field_section_size: ?u64 = null,
     enable_connect_protocol: bool = false,
 };
 
 pub const DecodeOptions = struct {
-    max_field_section_size: ?usize = null,
+    max_field_section_size: ?u64 = null,
     enable_connect_protocol: bool = false,
 };
