+++ tests/integration/production_preset.zig
@@ -28,7 +28,7 @@ test "Client.Config.production + Server.Config.production drive a basic GET end 
     const client_preset = http3_zig.Client.Config.production;
     const server_preset = http3_zig.Server.Config.production;
     try std.testing.expectEqual(@as(?usize, 256), client_preset.max_concurrent_peer_streams);
-    try std.testing.expectEqual(@as(?usize, 16 * 1024), client_preset.max_field_section_size);
+    try std.testing.expectEqual(@as(?u64, 16 * 1024), client_preset.max_field_section_size);
     try std.testing.expectEqual(@as(?usize, 16 * 1024), client_preset.wt_max_buffered_bytes_per_stream);
     try std.testing.expectEqual(http3_zig.SessionBufferedStreamPolicy.reject, client_preset.buffered_stream_policy);
     try std.testing.expectEqual(@as(?usize, 4 * 1024 * 1024), client_preset.max_event_payload_bytes_per_drain);
